requirements.txt
pillow
datasets
peft